Fix wrong times being saved for dates
DateTime.now() caused date to be saved with
different time than intended. This led to omitting
the dates when comparing two with same day,
but different time.
Time is hardcoded to be default value for now.
Added logic to compare two dates with precision
of a day when getting dates from database.